add another malicious testcase
Sentence initial "a" in the gold data is aligned
with the sentence final "a" in the system data,
although the spans clearly do not overlap.
However, the gold data contains a MWT,
which works like a "bridge" joining the two
system MWTs into one multiword span
and allowing LCS to align it.

Maybe this is not intuitive, but I don't think it is a bug.
Let's cover it in the tests anyway as an interesting edge case,
eventhough it is not likely it will ever appear in the real data.