make the align_words() code more readable (I hope)
There are two main branches of the alignment code
(mentioned also in the Description of token matching):
A: Multi-word tokens (in gold, system or both)
B: No multi-word tokens

The original code was correct, but handled some MWTs
even in the "no multi-word tokens" branch B,
so it was difficult for me to follow the main idea of the algorithm.
The new code handles the possible
"initial non-multiword words sticking out of the multiword span"
within the branch A, within the function find_multiword_span().
This function now finds both the start and the end
of the multiword span, which seems quite intuitive.

Note that there can be at most one "sticking out" word
(it could happen only if the word was sticking out of the end
of the preceding MWT).
This was not obvious from the original code.
In the new code, we could add a comment or assertion after `gi += 1`:
`assert gi==len(gold_words) or gold_words[gi].span.start >= system_words[si].span.start`
(similarly for si and system_words),
but I am not sure if it makes the code more readable/maintainable.