add sys-space.conllu which breaks the current script
Spaces in forms could be present also in the gold data
and it is quite probable that some parsers will miss this
and produce a word boundary on the space.
The evaluation script should not fail with the error
"The concatenation of tokens in gold file and in system file differ!"
in such cases.